+++ Kuplung/kuplung/objects/RayPicking.hpp
@@ -29,7 +29,7 @@ public:
     std::vector<RayLine*> rayLines;
 
 private:
-    int sceneSelectedModelObject;
+    unsigned short int sceneSelectedModelObject;
 
     void pickModel(const glm::mat4& matrixProjection, const glm::mat4& matrixCamera, const std::vector<ModelFaceBase*>& meshModelFaces, std::unique_ptr<ObjectsManager> &managerObjects, const std::unique_ptr<KuplungApp::Utilities::Input::Controls> &managerControls);
     void pickVertex(const glm::mat4& matrixProjection, const glm::mat4& matrixCamera, const std::vector<ModelFaceBase*>& meshModelFaces, std::unique_ptr<ObjectsManager> &managerObjects, const std::unique_ptr<KuplungApp::Utilities::Input::Controls> &managerControls);
