+++ Kuplung/kuplung/settings/Settings.h
@@ -42,7 +42,7 @@ public:
   void saveRecentFilesImported(const std::vector<FBEntity>& recentFiles);
   std::vector<FBEntity> loadRecentFilesImported();
 
-  int OpenGL_MajorVersion, OpenGL_MinorVersion;
+  unsigned short int OpenGL_MajorVersion, OpenGL_MinorVersion;
   bool maybeGracefullApplicationQuit;
   std::function<void(std::string)> funcDoLog;
   std::string ApplicationConfigurationFolder, appVersion, currentFolder, newLineDelimiter, SettingsFile, UIFontFile;
